requirements.txt
python-dotenv>=1.0.0
duckdb>=1.1.0
pandas>=2.2.0
pypdf>=4.0.0
pdfplumber>=0.11.0
qdrant-client>=1.11.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn>=0.32.0
pytest>=8.0.0
rich>=13.0.0
openai>=1.50.0
